requirements.txt
PySide2
maturin
pillow
diskcache